Moves test_getEXIFData from required to optional
As exifread isn't a requirement but rather something optional for whoever needs
it, tests using it shouldn't be inside tests.py. Travis should be happy now!
